+++ core/src/com/heshus/game/manager/Score.java
@@ -2,7 +2,7 @@ package com.heshus.game.manager;
 
 public class Score implements Comparable<Score> {
     private String name;
-    private long score;
+    private double score;
 
     // Zero-argument constructor for Json
     public Score() {
@@ -10,7 +10,7 @@ public class Score implements Comparable<Score> {
         this.score = 0;
     }
 
-    public Score(String name, long score) {
+    public Score(String name, double score) {
         this.name = name;
         this.score = score;
     }
@@ -18,12 +18,12 @@ public class Score implements Comparable<Score> {
     public String getName() {return name;}
     public void setName(String name) {this.name = name;}
 
-    public long getScore() {return score;}
+    public double getScore() {return score;}
     public void setScore(int score) {this.score = score;}
 
     // This is used when sorting all the scores created in the game
     @Override
     public int compareTo(Score o) {
-        return Long.compare(this.score, o.score);
+        return Double.compare(this.score, o.score);
     }
 }
